created component to hold connection func and return button
