+++ src/components/ControlBoard/index.jsx
@@ -1,0 +1,9 @@
+import './styles.css'
+
+export default function ControlBoard({buttons}) {
+    return (
+        <>
+        <div>fdsafdsafdsa</div>
+        </>
+    )
+}
