+++ src/components/ControlBoard/styles.css
@@ -1,0 +1,3 @@
+div {
+    background-color: salmon;
+}
